requirements.txt
python-telegram-bot==21.0
httpx==0.27.0
Pillow==10.2.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
requests==2.31.0
